Add validation to ensure Hugo is available in PATH
This adds a check at the top of these script files to ensure that `hugo`
is available in the PATH prior to making any changes to the local
filesystem or interacting with Git.

+++ review_gh_pages.sh
@@ -6,6 +6,8 @@ then
     exit 1;
 fi
 
+command -v hugo >/dev/null 2>&1 || { echo >&2 "I require hugo but it's not available in the PATH. Aborting."; exit 1; }
+
 echo "Deleting old publication"
 rm -rf public
 mkdir public
